Added classes Fixed test_merge.py Moved read and write functions
