+++ src/confluence/io/__init__.py
@@ -1,0 +1,11 @@
+# from .excel import ExcelReader
+# from .excel import ExcelWriter
+# from .text import TextReader
+# from .text import TextWriter
+# from .JSON import JSONReader
+# from .JSON import JSONWriter
+# from .CSV import CSVReader
+# from .CSV import CSVWriter
+# from .pif import PifWriter
+from .read import read
+from .write import write
